python/pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "pixelsrc"
version = "0.2.0"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
description = "Semantic pixel art format and compiler"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
numpy = ["numpy>=1.20"]
images = ["Pillow>=9.0"]

[tool.maturin]
manifest-path = "../Cargo.toml"
features = ["python"]
module-name = "pixelsrc._native"
